Update Texts_nb.properties
Reviewing my older translation, I assume "Make is safe" means the VPK it produces will be marked as safe in VitaShell for example, so I changed the string to reflect that more accurately. If this isn't the case, revert to "Make er trygg" instead of "Lag trygg VPK".
Corrected some strings and translated new ones.

The context of "UNSAFE_NOT_ACCEPTED" is a bit unclear to me, I will update this if I see any better way to translate it. The way I translated it now was as if the user declined to install a game with extended permission, and the program aborts the operation.