quick fix for ftp disconnect
we dont care about the response from the server if we are disconnected, because when we are already disconnected it throws an exception because it tries to send a QUIT command to a invalidn socket connection
we need to monitor the ftp connection all the time to see the status. this will be implemented by another commit soon
also sometimes its not updating the gamelisttable on xerpis ftp server. i added an addional updateEntries() call and it seems to be better now