Fix some exceptions while updating the GameListTable
If you hit the refresh button severals times in a second it throws exceptions, because sometimes model2 is then empty and trying to remove a row ends up with an exception. so check with the method 'any()' if there are any rows stored in newRows, then remove all rows from model2 using setRowCount(0) instead of while (...) removeRow., because its shorter and smarter and then fill model2 with the rows of newRows. it is basically just updating the GameListTable