Improve dump detection, show compression method
Normal homebrew do have the ATTRIBUTE entry in param.sfo the value 0x8000 (32768) stored, so display them as homebrew

It is actually not the compression level, its the compression method. 8 is deflate. this is read from param.sfo, but it should be implemented to read it from the zip file header when the vpk is available. in the zip header are much more informations stored