Improve popup context menu on game entry list
Disable "Show in Explorer/Finder" when the game/homebrew is on the vita only. there is no vpk to show, except you dump it (which could be implemented in the future)
The game title and it in the popup context menu are now in blue in bold to highlight it,
Also change on the list ctrl header
---
if you dont like these changes, just revert them.
i have alot of other improvments locally in my repo, but most of them are untested or unfinished. also i dont really know kotlin and i hate the syntax so much, hence everything eats alot of time, so i dont know if i will continue to improve vitaorganizer, because there is so much to improve