Check for valid vpkFolder, check for valid TITLE_ID
If for any reason the vpkFolder path gets invalid (like vitaorganizer is on my NAS, and i access it from windows and os x. Paths are different and not valid on the other system) then this results in analyzing files forever.

Game id//TITLE_ID has to be a length of 9 characters or vitashell refuses to install. skip those file or automatically fix the title id in the future

Display STEP_DONE when the gamelisttable is filled with entries. it is a bit confusing displaying it before cause it needs some time to fill the gamelisttable and whiile that you see a empty gamelisttable